move no hash password check dbfunc test file
