+++ server/package.json
@@ -14,11 +14,14 @@
   "author": "jouwana",
   "license": "ISC",
   "devDependencies": {
+    "@types/bcrypt": "^5.0.0",
     "@types/cors": "^2.8.13",
     "@types/express": "^4.17.17",
+    "@types/jsonwebtoken": "^9.0.1",
     "@types/node": "^18.11.9",
     "@types/pg": "^8.6.6",
     "@types/uuid": "^9.0.0",
+    "jest": "^29.5.0",
     "tsc-watch": "^5.0.3",
     "typescript": "^4.9.3"
   },
@@ -28,7 +31,7 @@
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
-    "jsonwebtoken": "^8.5.1",
+    "jsonwebtoken": "^9.0.0",
     "knex": "^2.3.0",
     "mongoose": "^6.8.0",
     "pg": "^8.8.0",
